Add 'Total' tests to new file WhenGettingTotals
